Split internal code into multiple files
