+++ internal/postrs/bridge_test.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sync/errgroup"
-
-	"github.com/spacemeshos/post/config"
 )
 
 func Test_deviceMutex(t *testing.T) {
@@ -42,18 +40,4 @@ func Test_deviceMutex(t *testing.T) {
 	lock1A.Unlock()
 
 	require.NoError(t, eg.Wait())
-}
-
-func TestTranslateScryptParams(t *testing.T) {
-	params := config.ScryptParams{
-		N: 1 << (15 + 1),
-		R: 1 << 5,
-		P: 1 << 1,
-	}
-
-	cParams := translateScryptParams(params)
-
-	require.EqualValues(t, 15, cParams.nfactor)
-	require.EqualValues(t, 5, cParams.rfactor)
-	require.EqualValues(t, 1, cParams.pfactor)
 }
